fix: popup test
